Refactor: 반복되는 formatData 함수 리팩토링 및 utils 폴더에 이동
